+++ airbyte-workers/src/main/java/io/airbyte/workers/internal/AirbyteMessageTracker.java
@@ -18,6 +18,8 @@ import io.airbyte.protocol.models.AirbyteRecordMessage;
 import io.airbyte.protocol.models.AirbyteStateMessage;
 import io.airbyte.protocol.models.AirbyteTraceMessage;
 import io.airbyte.workers.helper.FailureHelper;
+import io.airbyte.workers.internal.state_aggregator.DefaultStateAggregator;
+import io.airbyte.workers.internal.state_aggregator.StateAggregator;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
