+++ airbyte-workers/src/main/java/io/airbyte/workers/internal/state_aggregator/StateAggregator.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.internal;
+package io.airbyte.workers.internal.state_aggregator;
 
 import io.airbyte.config.State;
 import io.airbyte.protocol.models.AirbyteStateMessage;
