+++ airbyte-workers/src/test/java/io/airbyte/workers/internal/state_aggregator/StateAggregatorTest.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2022 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.workers.internal;
+package io.airbyte.workers.internal.state_aggregator;
 
 import static io.airbyte.protocol.models.AirbyteStateMessage.AirbyteStateType.GLOBAL;
 import static io.airbyte.protocol.models.AirbyteStateMessage.AirbyteStateType.LEGACY;
